Move helpers in utils and specific folder
